Refactoring library functions to lib
